Added project dependencies in requirements.txt
These libraries include:

requests: This is a Python library for sending HTTP requests. It abstracts the complexities of making requests behind a beautiful, simple API so that you can focus on interacting with services and consuming data in your application.

beautifulsoup4: This is a library for pulling data out of HTML and XML files. It works with your favorite parser to provide idiomatic ways of navigating, searching, and modifying the parse tree.

stem: This is a Python controller library that allows applications to interact with the Tor process.

PySocks: This is a SOCKS proxy module for Python. We'll use it to direct our requests through the Tor network.